Change min and max values to doubles
The integer type was previously used for the min and max values of the
Slider on Windows.
This was however changed in the delivery of:
https://github.com/callstack/react-native-slider/pull/259/files
but was partially inverted as the result of rebasing the conflicting
changes.
As the int->double change is good to have and should be kept, this
commit fixes the merge mistake by changing the integer type to double.

+++ src/windows/SliderWindows/SliderView.h
@@ -37,7 +37,7 @@ namespace winrt::SliderWindows::implementation {
 
         const double CalculateStepFrequencyPercentageValue(const double& stepPropertyValue) const noexcept;
 
-        int64_t m_maxValue, m_minValue;
+        double m_maxValue, m_minValue;
         double m_value;
     };
 }
